# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Revolution Network API",
    "version": "1.0.0",
    "description": "Comprehensive API for the Revolution Network platform - A crowdfunding and revolutionary content platform powered by The Anthony Letters",
    "contact": {
      "name": "Revolution Network Support",
      "email": "support@revolutionnetwork.com",
      "url": "https://revolutionnetwork.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://revolutionnetwork.com/api",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "User authentication and session management"
    },
    {
      "name": "Users",
      "description": "User management and profiles"
    },
    {
      "name": "Projects",
      "description": "Project creation and management"
    },
    {
      "name": "Donations",
      "description": "Payment and donation processing"
    },
    {
      "name": "Letters",
      "description": "The Anthony Letters system"
    },
    {
      "name": "Collaboration",
      "description": "Real-time collaboration features"
    },
    {
      "name": "Stripe",
      "description": "Payment processing and subscriptions"
    },
    {
      "name": "Admin",
      "description": "Administrative operations"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "userType": {
            "type": "string",
            "enum": ["creator", "supporter"],
            "description": "Type of user account"
          },
          "letterProgress": {
            "$ref": "#/components/schemas/LetterProgress"
          },
          "stats": {
            "$ref": "#/components/schemas/UserStats"
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "goalAmount": {
            "type": "number",
            "format": "float"
          },
          "currentAmount": {
            "type": "number",
            "format": "float"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "funded", "completed"]
          },
          "creatorId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Letter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookNumber": {
            "type": "integer"
          },
          "letterNumber": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "assignment": {
            "type": "string"
          }
        }
      },
      "LetterProgress": {
        "type": "object",
        "properties": {
          "completedLetters": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "currentBook": {
            "type": "integer"
          },
          "currentLetter": {
            "type": "integer"
          },
          "totalProgress": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "totalDonations": {
            "type": "number"
          },
          "totalProjects": {
            "type": "integer"
          },
          "totalLettersRead": {
            "type": "integer"
          }
        }
      },
      "APIKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "The API key (only shown once)"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rateLimit": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/user/profile": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user profile",
        "security": [{"bearerAuth": []}, {"apiKey": []}],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/letters": {
      "get": {
        "tags": ["Letters"],
        "summary": "Get all letters",
        "security": [{"bearerAuth": []}, {"apiKey": []}],
        "responses": {
          "200": {
            "description": "List of letters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "letters": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Letter"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/letters/{id}": {
      "get": {
        "tags": ["Letters"],
        "summary": "Get specific letter",
        "security": [{"bearerAuth": []}, {"apiKey": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Letter details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Letter"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List all projects",
        "security": [{"apiKey": []}],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "active", "funded", "completed"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/developers/keys": {
      "get": {
        "tags": ["Admin"],
        "summary": "List API keys",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/APIKey"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create new API key",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "rateLimit": {
                    "type": "integer",
                    "default": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKey"
                }
              }
            }
          }
        }
      }
    }
  }
}